Update patch set 2
Patch Set 2: Code-Review-1

It looks like all that has been added in the case of InfluxDB repository is an extension to the interface adding createdSince. However, it doesn't appear to be used anywhere so the implementation isn't complete. Am I misreading this? Also, I'm wondering how this will actually work in InfluxDB as we don't create definitions like we do in Vertica. So, there isn't even an immediate implementation that comes to mind.

My assumption is that we are checking in working implementations when we are submitting git reviews. If we want to commit incomplete code that can be done on branches prior to submitting a review.

Patch-set: 2
Reviewer: Gerrit User 2419 <2419@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1