Update patch set 1
Patch Set 1: Code-Review-1

Wondering if the query on line #54 needs to be changed.
I tried this on a metric name that had two different sets of dimensions.
Without the limit 1 on the end, the correct list comes back.  With limit 1 on the end I only get one set of dimensions.

Patch-set: 1