Update patch set 1
Patch Set 1:

The query on line 54 is only used to return the entire set of dimensions.  The query that retrieves the metric names with the unique set of dimensions is on line 107 (the line that was changed).  Previously, line 107 was using a regex which would result in metric names with the same prefix as the desired metric name being returned.  The query has been changed to a list of metric names which results in the exact set of metric names desired.  This has been tested on mini-mon.

Patch-set: 1
Reviewer: Gerrit User 12512 <12512@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0