Update patch set 1
Patch Set 1:

Derrick, I agree that this is not optimal because it is not obvious the vertica_jdbc.jar is not needed if using influxdb, but we are going to be reworking the start scripts and maybe we can find a better way of doing this. I really don't like the idea of the chef cookbook rewriting the start script based on whether we are using vertica. If the cookbook writes the whole script, then that would be a good solution

Patch-set: 1
Reviewer: Gerrit User 11809 <11809@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0