Update patch set 5
Patch Set 5:

Cindy, please see the following:

This is from python coding guideline. Even though underscore is allowed, by it is discouraged as indicated in the guideline. There is no benefit of making it monasca_api. In the project, you already have api package. thinking about that. monasca_api.api.something sounds really silly.

Package and Module Names

Modules should have short, all-lowercase names. Underscores can be used in the module name if it improves readability. Python packages should also have short, all-lowercase names, although the use of underscores is discouraged.

Since module names are mapped to file names, and some file systems are case insensitive and truncate long names, it is important that module names be chosen to be fairly short -- this won't be a problem on Unix, but it may be a problem when the code is transported to older Mac or Windows versions, or DOS.

When an extension module written in C or C++ has an accompanying Python module that provides a higher level (e.g. more object oriented) interface, the C/C++ module has a leading underscore (e.g. _socket).

Patch-set: 5