Update patch set 5
Patch Set 5:

I've reviewed the code in monasca/middleware/metric_validator.py and understand that this is one of the "filters" that can be specified.

I'm just wondering how you see this working for all of the other REST endpoints defined in the Monasca API, such as statistics, measurements, alarms, notification-methods, ...

There is a big

        if (env.get('PATH_INFO', '').startswith('/v2.0/metrics') and
                env.get('REQUEST_METHOD', '') == 'POST'):

at the top of in the filter. So, I'm assuming that this filter is in the list of filters for all the REST endpoints in Monasca. The if block at the top basically tests to see if this is for a POST to the /v2.0/metrics endpoint. 

Are you thinking that there will be a "validator" class defined for each endpoint and that each implementation of that endpoint implements a if block at the top to filter it down?

Patch-set: 5