Update patch set 5
Patch Set 5:

Not sure you saw this earlier, but I had some problems getting the MetricsValidator invoked, but my monasca.ini file looks correct. See above. Any suggestions?

Not sure I buy into the point of doing things this way yet. One of my concerns is around performance. In the case that the MetricsValidator is one of the filters, when the validator is invoked it has to do a number of things:

1. First it gets invoked meaning that it is an extra filter within the list of filters. That adds the overhead of a function call. 
2. If I understand your comments, depending on how many filters there are, such as events and metrics validators, each validator is invoked, and each filter, must do a check on the endpoint, such as if "/v2.0/metrics" to test whether it should be used. So if we had an events and a metrics endpoint, there are two filters. But we have a lot more than two endpoints.
3. Most importantly, the validator reads the request, validates it, and then reformats it for the downstream kafka dispatcher. The kafka dispatcher on_post method then reads the message and write it to kafka. This data flow should be compared to the case of just invoking a single on_post method that does the validation and publishing to the Kafka queue. In that case the message is only read once. Is there a way to avoid reading the data twice. Seems like we are penalizing the case when input validation is done.

I haven't measured the impact of all the above yet, but I'm imagining there is one.

My other comment is that doing things like this in filters doesn't really seem like it was the intended use of filters and it is leading to a lot more complicated code. In the normal use of filters that I'm accustomed to a filter applies to cross-cutting concerns, like authentication, logging, auditing, performance analysis, ... I have no problems using filters for this.

However, when we start using filters for the main-line logic, and it is different logic for every endpoint it is going to lead to other problems, like the performance problems mentioned above, but also it is more complicated code. For sure the intended usage of Monasca seems like a second class citizen in this structure.

From an organizational view it seems a lot clearer to me to have a method called "on_post" in a class called "MetricResource". This is how the Falcon API examples are structured, such as http://falcon.readthedocs.org/en/latest/user/tutorial.html#first-steps. If we had a class called MetricResource with on_post the method looks like

def on_post(req, res):
    msg = read_request(req)

    if validation_enabled:
        is not is_valid(msg):
            return...

    parsed_metric = parse_metric(req)
    send_metric(parsed_metric)

To handle the case HP's or IBM's case one way to address that is to use the GOF Template Method pattern. Define a base class called MetricResource with methods read_request, is_valid and parse_metric, and then derive HpMetricResource and IbmMetricResource which implement those methods in the ways in which we each want them implemented.

For example, HP's implementation of is is_valid would actually do the validation of the request, and IBM's wouldn't. HP's implementation of parse_metric would return the metric in it's envelope for downstream processing, and IBM's would just return the request body that is passed in. send_metric would be the same in both cases.

it is also unclear what IBM is designing for if there are messages that are unathenticated, not validated, not using the same message formats, ...

Do we need to have a call on this. Maybe we should get the teams together to discuss further.

Patch-set: 5