Update patch set 5
Patch Set 5:

My question regarding possible reasons why I can't get the metric validator invoked hasn't been addressed. Please seem my previous comments as I'm trying to test the code. Any suggestions on what is wrong?

I understand the use of validators in OpenStack. In all the examples you've provided I think they are a good use such as authentication, health-checks, rate-limits, ... I stated this in my previous post that I'm completely OK with validators in these case. These are cross-cutting concerns that are widely applicable to a broad number of methods.

The problem that I'm having is the use of validators for input validation and message formatting in the context that you are using it here. The input validation for a metric is not cross-cutting and broadly applicable. It applies to exactly one operation, POST /v2.0/metrics. It even does a test for the endpoint at the top. Similarly, this same logic would apply to event and any other endpoint. So, that seems like a pretty relevant difference from the examples of using auth, performance, rate-limiting, ..., for filters and a filter that only validates a metric request body and reformats it.

I completely realize you have to read the message at least once. My concern is that it looks like the combination of the metric validator and on_post method needs to read the message twice. Once in the metric validator in which it reads the message, parses it, and then stores the formatted message for the downstream processing. The other time for the on_post. Did I miss something? Is there a way to do this by only reading the message once. I was trying to run the code to validate this and analyze performance, but as I pointed out I didn't get the metric validator working. Looking at the code that seems to be what is being done. If that isn't the case then let me know. It just looks to me that there could be some extra over-head involved and I wanted to measure it. I don't want to penalize what I would consider to the the main-stream path for us. That path includes authentication, input validation, wrapping the metric in a envelope that include the tenant ID and some other info, and then sending to Kafka.

I also completely agree with the design right now for authentication and that this is done via the middleware and good separation of concerns. No problems in that area. That is exactly how we've done it in the Java implementation. Authentication is broadly applicable to a number of areas. 

Do you know of examples in OpenStack in which validators are being similar to how you are using them here?

Patch-set: 5