Restrict changes to existing Alarm Definitions
Once an Alarm Definition has been created, match_by and any metrics
in the expression cannot be changed. This is because those fields
control the metrics used to create Alarms and Alarms may already
have been created. The function, operator, period, periods and any
boolean operators can change, but not the metrics in subexpressions
or of the number sub expressions.  All other fields in an Alarm
Definition can be changed

Add tests for new functionality and some old for patch and update

Fix a bug where match-by was being cleared in patch if not given

Update api docs for the restrictions

Fix minor typos in api docs

Change-Id: I89a46838847b9cb73a073f9dd91db8f9391b6019