Distribute messages better among the Kafka partitions
Ensure that the same Metric always goes to the same partition. This
is done by using the tenant id, metric name and dimensions as the
key used to hash to a partition. This works as long as the number
of partitions does not change.

Distribute the events randomly among partitions by using a message
count as the key.

Add in a special kafka partitioner class that gives better distribution
for the metrics although I am not sure exactly why

Change-Id: Ia2ada1957198b9072fa957bc2a41791457664241