Update patch set 1
Patch Set 1:

If you look at the partitioner, I take the hash code and add it to the hash code shifted over 6 bits and 12 bits and then mod that with the number of partitions. Adding in those extra bits shouldn't make much difference, but it does. My only guess is that it has to do with the fact our tenant ids, metric names, and dimensions are ascii

Patch-set: 1
Label: Verified=0