Update patch set 1
Patch Set 1:

Is there a valid user that can be added to the roles monasca.conf file so that the current code can be made to work, or does the check in the code always fail? 

Deklan is basically spending valuable time trying to test your code and wondering why it is failing and how to get it to work.

Patch-set: 1
Label: Verified=0