Update patch set 1
Patch Set 1:

Hi David, I -1 this change because I don't think this is correct. The "value" field in a metric is not optional and it should be a int or float, but it can't be a string. Unless you are thinking that it could be a string with numeric values in it.

Patch-set: 1