Converted the three sql statements to one sql
This was done for better performance.Removed dead code that was no longer being used after this change.
Cleaned up the sql to only pick up necessary columns. Added a static inner class for a new mapper. Removed
unused imports and blank lines. Removed the unnecessary joins with the notification_method tables and tested
to make sure the functionality works

Change-Id: I314bc746cf25152db018e27c5a7f64ee7a0d3055