Update patch set 1
Patch Set 1:

Hi Victor, Thanks for submitting these, it really shows what you are thinking. I have been struggling with how to respond to these changes since you submitted them. It is just hard to have a real conversation in a review.  I talked to Roland today and we think it would be good to talk about them in the weekly Openstack meeting next Tuesday. We can talk about Ansible and some of the issues we have had with names, locations, etc. and hear about other people's opinions on these matters. Does that work for you?

Patch-set: 1