Add measurement valueMeta
valueMeta for measurements that have it will be sent in the message
to Kakfa. valueMeta for measurements that have it will be returned
when the measurement is read

Supported only by Influx V8 for now

Requires the changes to monasca-common from
https://review.openstack.org/160654

Removed ability to send arrays of timestamps and values. It was
undocumented and unused

Change-Id: Ifd7a64559abb17c1261d8fa85ece0d4548f93a8d
Implements: blueprint measurement-meta-data