Update patch set 28
Patch Set 28: Workflow-1

(2 comments)

This breaks the CLI and the smoke test because of the new format of the returned objects for the list methods. Instead of arrays, it now returns a json object containing the array and a the links. I know this is what is intended, we just have to update the CLI and merge it at the same time.

We also need to update monasca-ansible-api to add the new limit parameter to api-config.yml

Patch-set: 28
Reviewer: Gerrit User 11809 <11809@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Workflow=-1