Update patch set 7
Patch Set 7:

It's a great question Fabio.  I don't think that would work without additional code changes on the monasca side, similar to this patch.  The current code extracts the tenant_id from the token, to do sql queries that are tenant specific.  The underlying query code is all specific to a given tenant -- so even if the token were scoped to a domain, i think you'd still need the query parm for which tenant to get metrics for?

Patch-set: 7
Label: Verified=0