Update patch set 1
Patch Set 1:

(1 comment)

Hi Lukasz, Thanks for this significant contribution. In general, this all looks pretty good. My only significant comment is on whether we can avoid some code duplication.

InfluxV9AlarmStateHistoryHibernateRepo and AlarmStateHistoryVerticaRepoHibernateImpl seem to be almost duplicated classes. The cause of the duplication seems to be based on the class MySQLUtils which doesn't have an interface that it derives from. HibernateUtils was copied from MySQLUtils. So, in this case in the InfluxV9 and Vertica specific code it requires the MySQLUtils class in a few cases. I would like to try to avoid all the duplication. Could we define a new interface, Utils, then implement MySQLUtils and HibernateUtils by deriving from Utils. Then, instead of creating  InfluxV9AlarmStateHistoryHibernateRepo and AlarmStateHistoryVerticaRepoHibernateImpl, just modify the class that was there before, but instead of declaring InfluxV9AlarmStateHistoryHibernateRepo(HibernateUtils hibernateUtils, ..., and similar, the constructor for the previous code would be change to take the interface class Utils. For example, 

  public InfluxV9AlarmStateHistoryRepo(@Named("mysql") DBI mysql,
                                       MySQLUtils mySQLUtils,

would change to 

   public InfluxV9AlarmStateHistoryRepo(Utils, ...)

and similar for the Vertical implementation.

MySQLUtils or HibernateUtils would be plugged in at the config level.

I haven't had a chance to test everything in these changes. Hopefully, I'll get to that tomorrow. It looks like everything should work and there isn't much risk being added to the existing code. My only other concern is that a couple of dependencies were added to pom.xml. So, we'll need to be really careful that we don't break our internal build as we mirror the jars locally and don't install from external sources. I was also wondering if there was an easy way to plug-in MySQL or Hibernate, without having to have the dependencies required.

I was wondering if you ever tried the existing MySQL queries on POSTgres by just modifying the existing mysql connector to be a postgres connector instead.

Also wondering if you've ever tested the new Hibernate code with MySQL?

Patch-set: 1
Label: Workflow=0