Adds adminUserDomainName and adminProjectDomainName support to monasca-api
This change adds two additional optional configuration entries to the api-config.yml to
configure the monasca-api: adminUserDomainName and adminProjectDomainName. both are used
to scope the username or projectname with a domainname in case the monasca-api admin user
is not defined in the Default domain. Another change will follow, adding the same entries
for the corresponding ids instead of the names.

Change-Id: Ide981dc079be7ec18abfaadc456c5b24c386dfc5
Closes-Bug: #1484594