Update patch set 4
Patch Set 4:

Thanks for the reviews Roland and Deklan, and sorry for the confusion of thinking offset was currently not working aside from this patch.  Early versions of this timeslice change did have a problem with offset due to the way timeslice rounds and returns the start time stamp for each respective timeslice, but with PS4, this is fixed.

I think patch is good, offset is working and it's dropping my query time from 17s to around 5...

Patch-set: 4