Update patch set 4
Patch Set 4: Code-Review-1

Hi Witold,
Are you using postgres as your backend database? If so, you need to increase the size of your database tables to allow the number of characters not bytes in each field. In Mysqldb, the field lengths are in characters so it really can store 255 UTF-8 characters. From what I remember working with postgres in the past, it uses bytes instead of characters for the length so this code would be required. However, the spec says 255 characters not bytes, so if the postgres database can't handle 255 characters, the postgres field lengths need to be increased. We really don't want database specific code at this level in the API.

Patch-set: 4
Reviewer: Gerrit User 11809 <11809@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1