Update patch set 1
Patch Set 1: Code-Review-1

What is the purpose of this change?  The right way to change the host and port would be in the paste file that is used to launch gunicorn not in the oslo config file.  Right now an example of host and port configuration is available in monasca-api/monasca_api/etc/api-config.ini.

If this is just for testing purposes would accepting command line arguments be better?

Patch-set: 1
Reviewer: Gerrit User 14273 <14273@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1