Update patch set 2
Patch Set 2:

Deklan, I'm getting a little concerned here with the process that is being used to get all this code merged in. I understand it is only us working on this right now, and we're trying to get things completed in-time for the Tokyo Summit, but we probably need to be soliciting review comments from others orgs too.

Would it be possible to do all the remaining DevStack work for Monasca as a single review, using amends? You can mark as WIP until everything is ready to go. Then we can get folks to review and +2.

Patch-set: 2
Reviewer: Gerrit User 2419 <2419@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0