Update patch set 3
Patch Set 3:

The other option is to not do an exact conversion and assume the worst case upper bound on the number of bytes per character, which is 4. So take the length of the string and multiply by 4. That probably isn't a great answer in this case, as it could be overly conservative.

Patch-set: 3