Update patch set 5
Patch Set 5:

I wanted to really be done with this change. However, the Monasca API, https://github.com/stackforge/monasca-api/blob/master/docs/monasca-api-spec.md#post-v20metrics, says that the meta data is limited to 2048 characters. What is implemented now is restricting to 2048/4 characters. That is a pretty big difference. It is certainly conservative, and will fix the bug. The problem is that the Agent is expecting to to be able to send meta_value that is 2048 characters in length, and if it is actually testing for that size now, it is incorrect, and the API will want to reject a lot of values that are sent by the Agent.

I think the real fix is that the Vertica schema needs to be adjusted to VARCHAR(2048*4), and the API should just check for 2048. 

What do you think?

Patch-set: 5
Reviewer: Gerrit User 2419 <2419@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0