Update patch set 1
Patch Set 1:

Hi Witold, I think we should just get this merged and then patch up remaining changes ourselves it that is OK with you. The developers making this change don't really work on the Monasca code. However, they have done a significant amount of work for us to address all the concerns. If they've missed a few, we should follow-up and add the necessary changes.

This was a huge amount of work on their part. Personally, I'm amazed everything went relatively smoothly.

Patch-set: 1