Update patch set 1
Patch Set 1:

Though looking at the code before we launch the simple http server we make a wsgi_app using an existing paste file that should have the host and port defined.  Is it possible to yank that information out of the wsgi_app object and use that to configure the simple http server?

Patch-set: 1