Update patch set 4
Patch Set 4:

Hi Ryan, It is really nice code, but I was wondering if rather than testing whether the notification name exists, by first trying to find it in the database, and then if it doesn't exist, continuing to create it, how about, declaring a constraint on the sql (is there one already), such that the sql query will fail if it doesn't exist, or only inserting one if not exists. One problem with the way that this is done here is I believe there is a race condition. For example, if you create two notifications at the same time with the same name, it is possible that they will both be created. Right? But it is a pretty slim chance that this would occur.

Patch-set: 4
Label: Verified=0