Update patch set 15
Patch Set 15:

Just wondering why you've commented out all of the Tempest tests, except for one and the plan for getting them working? Is this just a preliminary step?

Would it be possible to use the attributes on tests to ignore tests, rather than commenting out. All of the Tempest have an attribute, For example,

    @test.attr(type="gate")

I'm wondering if these can be used to accomplish essentially the same thing, but in a cleaner way.

Patch-set: 15
Label: Verified=0