Add optional start_time and end_time for metrics list
This is a very useful feature for many of the dashboards,
when not using the merge flag.  These dashboards begin
with a metric-list, and then do a statistics call for each
unique set of dimensions returned.  For many dashboards,
metrics are returned where no data is currently being
collected (deleted VMs, etc), causing many unnecessary queries
for no data -- not to mention an ugly dashboard with tons
of noise.

This enhancement will help both the current grafana implementation
as well as our 2.0 port.

This patches support the java implementations for vertica and
influxdb, as well as the python implementation of the api
for influxdb (vertica not currently supported in python).

Change-Id: I683f2a53aaf2d2ad8005dd1542883636086aad4a