Set SERVICE_HOST, enable cache, install key, tempest and monasca only
Set SERVICE_HOT=192.168.10.6. When OpenStack services register with
Keystone, this will register using the IP address that is accessible
from the host OS.

Enable the vagrant-cachier to speed-up subsequent deploys.

Disable all OpenStack services except for mysql, rabbit, key and
monasca. This will reduce the size of the VM and speed-up
deploys.

Change-Id: I791cce4dbf3a7f1c2297fb288886aae35ada4658