Update patch set 2
Patch Set 2:

Apologies for the (very) late response.

For the API, at the moment, PUT is only used for updates. On POST requests, there are fields which can be frequently left empty (ex. an empty list for undetermined actions) so it makes sense to allow the user to omit the field and automatically fill it with a default. In contrast it is more unexpected to receive a default value in a PUT request if a field is omitted, so it is better to require the exact input and notify the user if they didn't specify a value. Incidentally, we have a field in the PUT request ("actions_enabled") that is not allowed in the POST request which I will fix.

Also, could you remove the -2 Roland? This change shouldn't be in front of anything older now.

Patch-set: 2
Label: Verified=0