Update patch set 1
Patch Set 1:

This is to try and get our log files into the set of output from the devstack gate jobs.  It looks like the normal openstack services run using a special function in the devstack environment that automatically archives the log files.  We're running all the Monasca services using the standard system tools so this change is just linking our normal log file location to the screen archive location so our log files will show up on the first page of the gate job output.

Patch-set: 1