Update patch set 2
Patch Set 2:

I don't think post_test is a good place to do that.  None of the other services running in the gate use that hook to get the logs archived so I can't tell if the timing is correct for that.  Easiest way is just to test for the existence of the env variable.

Patch-set: 2