Update patch set 3
Patch Set 3:

Depending on how the gate is configured the other services don't run in screen but they still generate screen logs if the appropriate env variable is set.  All we're doing is performing the same linking manually that is available via utility functions in the devstack environment.  We could use the utility functions directly but that would require changing the way we run all of our services.

Patch-set: 3