Fix alarm-updated message for the python API
The tenantId field in the alarm-updated message for the python API
was named tenant_id. This caused a NullPointerException in the
Threshold Engine since tenantId wasn't set.

Also, populate the sub_alarms fields in the alarm-transitioned
message

The first test allows the smoke test to get farther. It now
fails trying to check alarm history.

Change-Id: I3cc507b03178410a47bc72f4859e41310abf411b