Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Hi Shinya, I noticed this issue the other day too, but my thought on it initially was that the tenant_id shouldn't be returned in the case of InfluxDB either. So, to me it looks like a bug in the monasca-api that needs to be resolved, not a problem with the test. The response object returned by Vertica/InfluxDB and Java/Python should match, very closely.

Craig and Ben have +1'd already, which makes me thing I'm probably wrong, but I thought I would check.

--ROoand

Patch-set: 1