Update patch set 14
Patch Set 14:

Brad, Please see the review in progress at, https://review.openstack.org/#/c/250109/, which checks to see if the start time is < the end time. Would it be possible to add the same check. There is a helper for the Python code that Shinya has added.

Regards --Roland

Patch-set: 14
Label: Verified=0