Update patch set 4
Patch Set 4:

Hi Ghanshyam, We have an issue in our tests that when we delete alarm definitions at the end of a test, we delete all the alarm definitions. If tests are being run in parallel, then one test can impact another test that is running at the same time.

Deklan is trying to work-around that at the moment by commenting out several of the tests so we can start passing the gate jobs. Then we can come up with a correct solution, and uncomment the tests again.

If you know how to run the tests serially, please let us know. If that is possible, we can back-out the tests that were commented.

Patch-set: 4