Update patch set 1
Patch Set 1:

Though I wonder if the license is an issue there.

What's the ultimate goal here?  Are we trying to validate than an email is added correctly or do we want to verify that an email works?  Should we add some sort of verification email that the user needs to receive before we accept an address?

Patch-set: 1