Update patch set 10
Patch Set 10:

group_by isn't grouping correctly in the Java API.

In the monasca-vagrant environement, when running the Python API on the command, 

monasca alarm-count --group-by=state,severity

both the state and severity columns are returned.

However, in the Java API, only the state column is returned. 

Java API appears to be only returning the first column.

Patch-set: 10
Label: Verified=0