Update patch set 1
Patch Set 1:

I wanted to add new tests for sort_by in change with orm. But unfortunately I found that with change "Fix error when sorting with limit" the issue was introduced. Issue is connected with replacing queery_parms['sort_by']. The details are in results of job gate-monasca-api-tempest-dsvm. And I really don't know to fix it quickly.

Patch-set: 1
Label: Verified=0