Update patch set 1
Patch Set 1:

Thank you for your +2. But now what should I do? Because part of this workaround stuff I made in https://review.openstack.org/#/c/266922/24 and change 266922 is parent of this change. Should I move these part of workaround to this change or do you want to merge both 282201 and 266922?

Patch-set: 1