Update patch set 1
Patch Set 1:

Not only by downgrading. In moansca-api in change with orm I tried to repair tests. Bu I found that openstack increase version of kafka-python in project requirments. And for that reason downgrading for monasca-api didn't work. So in change with orm I increase version again version of kafka-python to 1.0.0 and I change worker to gevent because I found that it could be better for kafka-python. But for monasca-notificatoin and monasca-persister I had to downgrade versions of kafka-python. And I did it in this change. And I'm supriesed that it's working. I think that you could consider merge also change with orm. And both changes together should give you working tests and support for orm :)).

Patch-set: 1