Update patch set 1
Patch Set 1:

Brad, This code will probably interest you as it adds support for returning multiple metrics in a single http request body in the measurements and statistics resources. 

The code works, except pagination doesn't work and InfluxDB is not supported. I'm not sure how to resolve that yet, but I just wanted to test the code and make it available for performance analysis and feedback.

There were no changes required to the python-monascaclient. However, you might need to modify Grafana.

Patch-set: 1