Update patch set 4
Patch Set 4:

Thanks Brad. We'll look at the id issue. Thanks for testing.

Ryan and I discussed this today, and are going to make a change to future-proof the api. 

Instead of the query parameter "multiple_metrics", we will switch to "group_by". Initially, the group_by will support one value of "*". It will look like the following:

group_by=*

Eventually, we will allow you to specify dimensions keys to the group_by clause.

The  reason for the change is that the multiple_metrics flag i was adding is really a group_by all. So, there is no reason to have a special case query parameter "multiple_metrics" just for this one case. 

group_by=*

is the same as multiple_metrics=true

Then in the future you'll be abel to do

group_by=region 

or 

group_by=zone

or

group_by=region,zone

where region and zone are dimension keys.

Patch-set: 4