Update patch set 6
Patch Set 6:

Rohit, The goal will be to add the group_by, but initially it will just be group_by all. However, we'll follow-up adding the full group_by. 

This still needs to be reviewed by the community and approved., but I'm assuming it will get adopted, unless someone comes up with some good reasons why it shouldn't be adopted or a better idea.

Patch-set: 6
Label: Verified=0