Update patch set 4
Patch Set 4: Code-Review+1

Hi Shinya, This all looks good to me. Thanks for adding this!

I would prefer merging this in a week, after the final Mitaka release is tagged as I'm trying not to introduce any new functionality at this stage in the release cycle. I should be tagging by the end of the week.

We should get a couple more reviewers to review this.

We'll need to check with Witold on the Java/Hibernate support as that isn't implemented again. I think Witold was in the process of adding support for the other query parameters that we recently added.

Patch-set: 4
Label: Code-Review=+1