Update patch set 6
Patch Set 6:

I guess that is something I've been thinking ona adding to InfluxV9Utils, but Deklan explained it to me that it wouldn't really be the case here. 

As far as I remember it was explained here: https://review.openstack.org/#/c/226733/.

Check comment on PS9.

Patch-set: 6
Label: Verified=0