Update patch set 6
Patch Set 6:

Looks fine to me, except for possible name change to "ignore_undetermined" or "no_undetermined", ...

Also, we need to discuss the "latched" attribute, and that one probably needs a better name too. This review and the related one's prevent the alarm going to the UNDETERMINED state, but don't address keeping it in the ALARM'd state until manually set.

Patch-set: 6
Label: Verified=0