Update patch set 15
Patch Set 15:

There has been comments on the review https://review.openstack.org/#/c/296813/11 that we should switch the period to seconds from minutes. I thought this review would be a better place to have the discussion as it is where the field is documented and defined. 

The reason why is was brought up, by Roland, was because if we move to seconds it would give us the ability in the future to move to second granularity including sub minute periods easily. In his words we would be 'future proof'. If we stay with it being in minutes that would restrict us from doing that without changing the definition of the field and breaking behavior in the API. By going with Roland's suggestion we would be a lot more flexible moving forward. 

I agree with Roland but I wanted to get others' opinions on this before making the change from minutes to seconds. 

What do other people think?

Another thing is we decided on the field name to be 'period' instead of periodic_interval or interval. If anyone has a problem on that please let us know.

Patch-set: 15