Update patch set 15
Patch Set 15:

One of my concerns with a sub-minute period is that we don't know how long a webhook may take to return.  If we allow periods under a minute we could more easily run into issues maintaining a consistent interval between sends on a give webhook.

This potential problem still exists with a one minute period so I've asked the Heat team how many scale groups they expect to have and how long it might take to respond to a webhook.

This is fixable on the notification side with some additional configuration parameters allowing the user to run multiple periodic engines for a given period.

Patch-set: 15
Label: Verified=0