Update patch set 4
Patch Set 4:

One other potential suggestion, but if you want to stick with enabling the hint in the yaml/config file then rather than "provideDbHint" you might want to change to "provideVerticaReplicationHint" to make it really clear that this is a Vertica only hint and that it is for telling the query optimizer that the data is fully replicated. 

There might be a better name than what I've come up with that matches what Vertica calls this in the documentation, but the main idea is that the DbHint is a Vertica specific query that is informing the query optimizer that the data is fully replicated and therefore local.

Is there a reference in the documentation on this hint. I couldn't find it.

Patch-set: 4