Update patch set 7
Patch Set 7:

thanks for the review roland!  i like the idea of actually providing the hint in the yaml file, making more extensible in the future if we think of other hints that make sense.  in order to achieve horizontal scaling at twc, we have replicated all tables, including measurements.  we knew this would be a hit in terms of write speed, but it allows any node in the cluster to satisfy any metrics related query.  in terms of the documentation for this hint, you are correct, it's undocumented.  it was  a suggestion from the vertica engineers when here onsite.  

i'll push another patch shortly -- thanks!

Patch-set: 7
Label: Verified=0