Update patch set 1
Patch Set 1: Code-Review+1

Hi Shinya, thank you for finding this. It took me a long time though to determine why this was an issue. Perhaps next time you could put more detail in the commit message.

For other people reviewing this, this query is only called from alarms_repository.get_alarm_definition which is only called from alarms.py in _alarm_update and _alarm_patch when the state of an alarm is being changed. In that case, the code is trying to get the Alarm Definition using the alarm id. Unfortunately, the code that calls it have clauses like:

            except exceptions.DoesNotExistException:
                # Alarm definition does not exist. May have been deleted
                # in another transaction. In that case, all associated
                # alarms were also deleted, so don't send transition events.
                pass

so the fact the query failed was never detected. This results in no message being sent to the AlarmStateTransition topic that the state has been changed by the API and thus the alarm history is not saved by the persister. You can see this by using monasca alarm-history.

Would be good if we had a tempest test for this particular case.

Patch-set: 1
Reviewer: Gerrit User 11809 <11809@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1